[feat] [PULSE-274] alert component
