--- src/features/mentor/popupSearch/index.ts
@@ -1,3 +1,0 @@
-export * from "./FieldPopup";
-export * from "./LearningTypePopup";
-export * from "./RegionPopup";
